Ensure we match all known overloads
